bp_moonbeam
Trait FromMoonbeamPolkadotInboundLaneApi

pub trait FromMoonbeamPolkadotInboundLaneApi<Block: BlockT>: Core<Block> {
    // Provided method
    fn message_details(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        lane: LaneId,
        messages: Vec<(MessagePayload, OutboundMessageDetails)>,
    ) -> Result<Vec<InboundMessageDetails>, ApiError> { ... }
}

Inbound message lane API for messages sent by this chain.

This API is implemented by runtimes that are receiving messages from this chain, not by this chain’s runtime itself.

Entries of the resulting vector are matching entries of the messages vector. Entries of the messages vector may (and need to) be read using To<ThisChain>OutboundLaneApi::message_details.

Provided Methods

fn message_details( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, lane: LaneId, messages: Vec<(MessagePayload, OutboundMessageDetails)>, ) -> Result<Vec<InboundMessageDetails>, ApiError>

Return details of given inbound messages.