bp_moonbeam

Trait ToMoonbeamPolkadotOutboundLaneApi

pub trait ToMoonbeamPolkadotOutboundLaneApi<Block: BlockT>: Core<Block> {
    // Provided method
    fn message_details(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        lane: LaneId,
        begin: MessageNonce,
        end: MessageNonce,
    ) -> Result<Vec<OutboundMessageDetails>, ApiError> { ... }
}

Outbound message lane API for messages that are sent to this chain.

This API is implemented by runtimes that are receiving messages from this chain, not by this chain’s runtime itself.

Provided Methods

fn message_details( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, lane: LaneId, begin: MessageNonce, end: MessageNonce, ) -> Result<Vec<OutboundMessageDetails>, ApiError>

Returns dispatch weight, encoded payload size and delivery+dispatch fee of all messages in given inclusive range.

If some (or all) messages are missing from the storage, they’ll also will be missing from the resulting vector. The vector is ordered by the nonce.